% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_cancer}
\alias{atlas_cancer}
\title{32 cancer types monopoly summary}
\format{
data.frame: cancer, n_tumor, n_mono, mono_rate, tmi5_median, top_carriers
}
\usage{
atlas_cancer
}
\description{
32 cancer types monopoly summary
}
\keyword{datasets}
