% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_demand}
\alias{atlas_demand}
\title{74 genes x 52 GTEx tissues demand (expression percentile, long)}
\format{
data.frame: gene, tissue, pctile, med_tpm
}
\usage{
atlas_demand
}
\description{
74 genes x 52 GTEx tissues demand (expression percentile, long)
}
\keyword{datasets}
