% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_gene_cancer}
\alias{atlas_gene_cancer}
\title{74 genes x 32 cancers monopoly frequency (long)}
\format{
data.frame: gene, cancer, mono_freq, n_mono, n_mono_samples
}
\usage{
atlas_gene_cancer
}
\description{
74 genes x 32 cancers monopoly frequency (long)
}
\keyword{datasets}
