% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_genes}
\alias{atlas_genes}
\title{74 monopoly genes annotation (36 cancer-specific + 38 tissue-shared)}
\format{
data.frame: gene, group, module, family, secreted_flag, competence scores
}
\usage{
atlas_genes
}
\description{
74 monopoly genes annotation (36 cancer-specific + 38 tissue-shared)
}
\keyword{datasets}
