% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_top_monopoly}
\alias{get_top_monopoly}
\title{Top monopoly genes by cancer occurrence}
\usage{
get_top_monopoly(n = 10, gc = NULL)
}
\arguments{
\item{n}{number}

\item{gc}{optional data.frame (默认=内置 atlas_gene_cancer)}
}
\value{
data.frame
}
\description{
Top monopoly genes by cancer occurrence
}
