% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{monopoly_palette}
\alias{monopoly_palette}
\title{monopoly palette (Okabe-Ito 扩展, 正文主色)}
\usage{
monopoly_palette(n = 8)
}
\arguments{
\item{n}{需要的颜色数 (1-8)}
}
\value{
character vector of hex colors
}
\description{
monopoly palette (Okabe-Ito 扩展, 正文主色)
}
