% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monopolyAtlas-package.R
\docType{package}
\name{monopolyAtlas-package}
\alias{monopolyAtlas}
\alias{monopolyAtlas-package}
\title{monopolyAtlas: Transcriptomic Monopoly Analysis Framework}
\description{
Compute TMI/top-share monopoly metrics, query the 74-gene pan-cancer monopoly
atlas, and reproduce the manuscript visual language (STATE/CARRIER/WHERE/
WHICH/MEANING/BOUNDARY).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/liufangyuan0613/monopolyAtlas}
  \item Report bugs at \url{https://github.com/liufangyuan0613/monopolyAtlas/issues}
}

}
\author{
\strong{Maintainer}: Fangyuan Liu \email{liufangyuan1989@126.com}

}
\keyword{internal}
