% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_tmi_ridge}
\alias{plot_tmi_ridge}
\title{TMI5 distribution by cohort (ridge/violin)}
\usage{
plot_tmi_ridge(df, order_by = c("median", "none"))
}
\arguments{
\item{df}{data.frame: sample, tmi5, cohort (cancer/tissue), state (mono/nonmono/normal 可选)}

\item{order_by}{"median" (默认) 或 "none"}
}
\value{
ggplot
}
\description{
TMI5 distribution by cohort (ridge/violin)
}
