% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gene}
\alias{query_gene}
\title{Query a gene in the monopoly atlas (v3: 74-gene 全注释)}
\usage{
query_gene(gene_symbol, genes_meta = NULL)
}
\arguments{
\item{gene_symbol}{gene symbol}

\item{genes_meta}{optional data.frame (默认=内置 atlas_genes)}
}
\value{
named list
}
\description{
Query a gene in the monopoly atlas (v3: 74-gene 全注释)
}
