% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_monopoly}
\alias{theme_monopoly}
\title{ggplot2 theme for monopoly figures}
\usage{
theme_monopoly(base_size = 12)
}
\arguments{
\item{base_size}{base font size}
}
\description{
ggplot2 theme for monopoly figures
}
